+++ internal/handler/web/health.go
@@ -1,0 +1,14 @@
+package web
+
+import "net/http"
+
+func HealthCheck(rw http.ResponseWriter, req *http.Request) {
+	if req.Method != http.MethodGet {
+		rw.WriteHeader(http.StatusMethodNotAllowed)
+		rw.Write([]byte(`{"error": "method not allowed"}`))
+		return
+	}
+
+	rw.WriteHeader(http.StatusOK)
+	rw.Write([]byte(`{"status": "OK"}`))
+}
