+++ internal/handler/web/middleware.go
@@ -1,0 +1,17 @@
+package web
+
+import (
+	"net/http"
+)
+
+func Middleware(h http.HandlerFunc) http.HandlerFunc {
+	return func(rw http.ResponseWriter, req *http.Request) {
+		rw.Header().Set("Access-Control-Allow-Origin", "*")
+		rw.Header().Set("Access-Control-Allow-Credentials", "true")
+		rw.Header().Set("Access-Control-Allow-Headers", "Content-Type, Content-Length, Accept-Encoding, X-CSRF-Token, Authorization, accept, origin, Cache-Control, X-Requested-With")
+		rw.Header().Set("Access-Control-Allow-Methods", "POST, OPTIONS, GET, PUT, PATCH, DELETE")
+		rw.Header().Set("Content-Type", "application-json")
+
+		h.ServeHTTP(rw, req)
+	}
+}
