+++ internal/handler/web/web.go
@@ -4,27 +4,9 @@ import (
 	"encoding/base64"
 	l "fiap-hf-src/pkg/logger"
 	"fmt"
-	"net/http"
 	"os"
 	"strings"
 )
-
-func HealthCheck(rw http.ResponseWriter, req *http.Request) {
-	rw.Header().Add("Access-Control-Allow-Origin", "*")
-	rw.Header().Add("Access-Control-Allow-Credentials", "true")
-	rw.Header().Add("Access-Control-Allow-Headers", "Content-Type, Content-Length, Accept-Encoding, X-CSRF-Token, Authorization, accept, origin, Cache-Control, X-Requested-With")
-	rw.Header().Add("Access-Control-Allow-Methods", "POST, OPTIONS, GET, PUT, PATCH, DELETE")
-	rw.Header().Add("Content-Type", "application-json")
-
-	if req.Method != http.MethodGet {
-		rw.WriteHeader(http.StatusMethodNotAllowed)
-		rw.Write([]byte(`{"error": "method not allowed"}`))
-		return
-	}
-
-	rw.WriteHeader(http.StatusOK)
-	rw.Write([]byte(`{"status": "OK"}`))
-}
 
 func tokenValidate(tokenInput string) error {
 	apiHTokenDecode, err := base64.StdEncoding.DecodeString(tokenInput)
